+++ vitest.config.ts
@@ -1,3 +1,10 @@
 import {defineConfig} from 'vitest/config'
 
-export default defineConfig({test: {environment: 'node', include: ['packages/**/__tests__/**/*.test.ts'], testTimeout: 30000}})
+export default defineConfig({
+  test: {
+    environment: 'node',
+    include: ['packages/**/__tests__/**/*.test.ts'],
+    setupFiles: ['./vitest.setup.ts'],
+    testTimeout: 30000,
+  },
+})